todo: lazy loading on the Shelters livewire component
